:bug: #1425 修复企业微信OA审批相关几个参数的类型或命名

+++ weixin-java-cp/src/main/java/me/chanjar/weixin/cp/bean/oa/applydata/ContentValue.java
@@ -11,16 +11,15 @@ import java.util.List;
  */
 @Data
 public class ContentValue implements Serializable {
-
   private static final long serialVersionUID = -5607678965965065261L;
 
   private String text;
 
   @SerializedName("new_number")
-  private Integer newNumber;
+  private Double newNumber;
 
   @SerializedName("new_money")
-  private Integer newMoney;
+  private Double newMoney;
 
   private ContentValue.Date date;
 
@@ -36,24 +35,21 @@ public class ContentValue implements Serializable {
 
   @Data
   public static class Date implements Serializable {
-
     private static final long serialVersionUID = -6181554080062231138L;
     private String type;
 
     @SerializedName("s_timestamp")
-    private Long timestamp;
+    private Double timestamp;
   }
 
   @Data
   public static class Selector implements Serializable {
-
     private static final long serialVersionUID = 7305458759126951773L;
     private String type;
     private List<Option> options;
 
     @Data
     public static class Option implements Serializable {
-
       private static final long serialVersionUID = -3471071106328280252L;
       private String key;
 
@@ -74,28 +70,23 @@ public class ContentValue implements Serializable {
 
   @Data
   public static class Department implements Serializable {
-
     private static final long serialVersionUID = -2513762192924826234L;
 
     @SerializedName("openapi_id")
-
     private String openApiId;
     private String name;
   }
 
   @Data
   public static class File implements Serializable {
-
     private static final long serialVersionUID = 3890971381800855142L;
+
     @SerializedName("file_id")
     private String fileId;
-
-
   }
 
   @Data
   public static class Child implements Serializable {
-
     private static final long serialVersionUID = -3500102073821161558L;
     private List<Content> list;
   }
